OTel.Instrumentation.Wai.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

buildMiddleware :: TracerProvider -> IO Middleware

middleware :: TracingBackend -> Middleware

spanSpecFromRequest :: Request -> SpanSpec

attrsFromRequest :: Request -> AttrsBuilder 'AttrsForSpan

includeReqLengthIfKnown :: Request -> Maybe (AttrsBuilder 'AttrsForSpan)

decodeBytes :: ByteString -> Text